TIP-886: Move UserManagement tests under tests/
