+++ src/Akeneo/UserManagement/.php_cd.php
@@ -5,8 +5,6 @@ use Akeneo\CouplingDetector\Configuration\DefaultFinder;
 use Akeneo\CouplingDetector\RuleBuilder;
 
 $finder = new DefaultFinder();
-$finder->notPath('spec');
-$finder->notPath('tests');
 
 $builder = new RuleBuilder();
